util/erq/erq.c: Fix buffer overrun
Replace use of long datatype with int32_t.  When this code was written,
longs were always 4 bytes.  These days, on a 64-bit system, a long is 8
bytes.  This causes a buffer overrun problem in a few places.

+++ src/util/erq/erq.c
@@ -155,7 +155,7 @@ union ticket_u
 {
     struct ticket_s
     {
-        long rnd, seq;
+        int32_t rnd, seq;
     } s;
     char c[1];
 };
